--- frontend/src/components/fileManager/types.ts
@@ -1,14 +1,0 @@
-import { FileWithUrl } from '../../types/file';
-
-export type FileSource = 'recent' | 'local' | 'drive';
-
-export interface FileListItemProps {
-  file: FileWithUrl;
-  isSelected: boolean;
-  isSupported: boolean;
-  onSelect: () => void;
-  onRemove: () => void;
-  onDoubleClick?: () => void;
-  isLast?: boolean;
-}
-
